+++ server/database/models/petRepository.js
@@ -1,18 +1,20 @@
 const AbstractRepository = require("./AbstractRepository");
 
-class petRepository extends AbstractRepository {
+class PetRepository extends AbstractRepository {
   constructor() {
     super({ table: "pet" });
   }
 
-  async readAll() {
-    const [rows] = await this.database.query(`select petName, pet_type.name as typeName, breed, age, information from ${this.table} join pet_type on type_id= pet_type.id`);
+  async readAll(userId) {
+    const [rows] = await this.database.query(`select pet.id, petName, pet_type.name as typeName, breed, age, information from ${this.table} join pet_type on type_id= pet_type.id where user_id=?`,
+    [userId]
+  );
     return rows;
   }
 
   async create(pet) {
     const [result] = await this.database.query(
-      `insert into ${this.table} ( petName, type_id, breed, age, information, user_id) values (?, ?, ?, ?, ?,?)`,
+      `insert into ${this.table} ( petName, type_id, breed, age, information, user_id) values (?, ?, ?, ?, ?, ?)`,
       [
         pet.petName,
         pet.type_id,
@@ -27,12 +29,12 @@ class petRepository extends AbstractRepository {
 
   async delete(pet) {
     const [result] = await this.database.query(
-      `DELETE FROM ${this.table} WHERE id = ?`,
-      [pet.id]
+      `DELETE FROM ${this.table} WHERE id = ? AND user_id = ?`,
+      [pet.id, pet.userId]
     );
 
     return result.affectedRows;
   }
 }
 
-module.exports = petRepository;
+module.exports = PetRepository;
